[Feat] Notice pagination, UserInfo Update token 인증

+++ src/main/java/com/homeis/user/controller/UserController.java
@@ -75,7 +75,14 @@ public class UserController {
 	}
 	
 	@PutMapping
-	public ResponseEntity<Integer> updateUserInfo(@ModelAttribute User userInfo) {
+	public ResponseEntity<?> updateUserInfo(@RequestBody User userInfo, @RequestHeader("Authorization") String tokenHeader) {
+		//토큰이 유효하지 않은 경우
+		boolean isValid = jwtUtil.isValid(tokenHeader.substring(7));
+		if(!isValid) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰 만료, 잘못된 접근.");
+		
+		String userId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		if(!userId.equals(userInfo.getId())) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("잘못된 접근입니다.");
+		
 		int flag = userService.updateUserInfo(userInfo);
 		if (flag == 0) {
 			return ResponseEntity.status(404).body(0);
@@ -83,8 +90,16 @@ public class UserController {
 		return ResponseEntity.status(200).body(1);
 	}
 	
-	@DeleteMapping("/{userId}")
-	public ResponseEntity<Integer> deleteUser(@PathVariable("userId") String userId) {
+	@DeleteMapping("/{id}")
+	public ResponseEntity<?> deleteUser(@PathVariable("id") String id, @RequestHeader("Authorization") String tokenHeader) {
+		
+		//토큰이 유효하지 않은 경우
+		boolean isValid = jwtUtil.isValid(tokenHeader.substring(7));
+		if(!isValid) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰 만료, 잘못된 접근.");
+		
+		String userId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		if(!userId.equals(id)) return ResponseEntity.status(HttpStatus.FORBIDDEN).body("잘못된 접근입니다.");
+		
 		int flag = userService.deleteUser(userId);
 		if (flag == 0) {
 			return ResponseEntity.status(404).body(0);
